Add functions to app.js
