+++ package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.18.2",
+    "bundle-js": "^1.0.2",
     "dotenv": "^5.0.1",
     "express": "^4.16.3",
     "pug": "^2.0.3",
